move to separate duplicates.py
